% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{calc_group_expr_num}
\alias{calc_group_expr_num}
\title{Calculate the Number of Expressed IsomiRs}
\usage{
calc_group_expr_num(x)

\S4method{calc_group_expr_num}{IsomirDataSet}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{IsomirDataSet} containing isomiR
expression data.}
}
\value{
A data frame with the following columns:
  \itemize{
    \item group: The name of the group.
    \item total: The total number of isomiRs expressed in the group.
    \item organ_specific: The number of isomiRs specific to the group
          (expressed in only this group).
    \item common: The number of isomiRs expressed in all groups.
  }
}
\description{
This method calculates the number of expressed isomiRs for each group in an
\linkS4class{IsomirDataSet} object.
It categorizes isomiRs into three types:
1. Total: The total number of isomiRs expressed in each group.
2. Organ-Specific: The number of isomiRs expressed in only one group.
3. Common: The number of isomiRs expressed in all groups.
}
\examples{
# Load example data
data(isomirs)

# Calculate the number of expressed isomiRs
calc_group_expr_num(isomirs)

}
