% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_one_cigar}
\alias{calc_one_cigar}
\title{Get CIGAR string represents alignment between query and ref}
\usage{
calc_one_cigar(query, ref)
}
\arguments{
\item{query}{query string}

\item{ref}{reference}
}
\value{
CIGAR
}
\description{
Get CIGAR string represents alignment between query and ref
}
