% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{calc_one_tsi}
\alias{calc_one_tsi}
\title{Calculate tissue specificity index of an expression profile.}
\usage{
calc_one_tsi(exp)
}
\arguments{
\item{exp}{Expression profile of a gene}
}
\value{
tau
}
\description{
\code{calculate_tsi} returns the tau value of a gene expression profile
  across tissues. The index \eqn{\tau} is defined as:
  \deqn{
   \tau=\frac{\sum_{i}^N (1-x_i)}/N-1
  }
}
