% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{calc_tpm}
\alias{calc_tpm}
\title{Calculate transcripts per million for each isoform}
\usage{
calc_tpm(isoforms)
}
\arguments{
\item{isoforms}{description}

\item{min_tpm}{description

\deqn{
 TPM = \frac{q_i/l_i}{\sum q_j/l_j} 10^6
}}
}
\description{
Calculate transcripts per million for each isoform
}
