% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{cluster_isoforms}
\alias{cluster_isoforms}
\title{Cluster Isoforms by Reference miRNAs}
\usage{
cluster_isoforms(group2isoforms)
}
\arguments{
\item{group2isoforms}{A list of data frames, where each data frame contains
isoform information for a specific group (e.g., tissue or treatment). Each
data frame must include the following columns:
\itemize{
  \item mature_id: Reference miRNA ID.
  \item mature_seq: Reference miRNA sequence.
  \item seed_seq: Seed sequence (center region of the mature miRNA).
  \item template_seq: Template sequence including flanking regions.
  \item read_seq: Read sequence matched to the reference.
  \item indel_5p: Number of indels at the 5' end.
  \item indel_3p: Number of indels at the 3' end.
  \item snp_5p: Number of SNPs at the 5' end.
  \item snp_3p: Number of SNPs at the 5' end.
  \item snp_seed:
  \item dist
}}
}
\value{
A list of \linkS4class{Isomir}.
}
\description{
This function clusters isoforms by their reference miRNAs and organizes the
data into a list of \linkS4class{Isomir} objects. It performs multiple
sequence alignment (MSA) using the provided sequences and generates CIGAR
strings to represent the alignment differences. It ensures that isoforms are
grouped by their reference miRNA and assigns unique identifiers to each
isoform based on its CIGAR string.
}
\details{
The function performs the following steps:
1. Extracts relevant columns (mature_id, mature_seq, template_seq, read_seq,
   indel_5p, indel_3p, snp_5p, snp_3p, snp_seed, dist) from each group's
   isoform data.
2. Combines all isoforms into a single data frame and removes duplicates.
3. Splits the isoforms by reference miRNA (mature_id).
4. Performs multiple sequence alignment (MSA) using the isoform, reference
   miRNA and template sequences, and generates CIGAR strings to represent
   the alignment differences.
5. For each reference miRNA, creates an \linkS4class{Isomir} object
   containing the associated isoforms.
}
