% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_isoform}
\alias{detect_isoform}
\title{Detect isoform based on the sample infomation file containing sample name,
tissue/treatment and FASTQ path}
\usage{
detect_isoform(
  sample_info_file,
  spe,
  word_size = 13,
  max_ed_5p = 2,
  max_ed_3p = 3,
  min_tpm = 2
)
}
\arguments{
\item{sample_info_file}{description}

\item{spe}{description}

\item{word_size}{description}

\item{max_ed_5p}{description}

\item{max_ed_3p}{description}

\item{min_tpm}{description}
}
\value{
description
}
\description{
Detect isoform based on the sample infomation file containing sample name,
tissue/treatment and FASTQ path
}
