% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_isomirs}
\alias{detect_isomirs}
\title{Detect isoform based on the sample information file containing sample name,
tissue/treatment and FASTQ path}
\usage{
detect_isomirs(
  sample_info_file,
  fq_dir,
  mirnas,
  max_ed_5p = 2,
  max_ed_3p = 3,
  min_tpm = 2,
  type = "all"
)
}
\arguments{
\item{sample_info_file}{description}

\item{fq_dir}{description}

\item{mirnas}{description}

\item{max_ed_5p}{description}

\item{max_ed_3p}{description}

\item{min_tpm}{description}
}
\value{
IsomirDataSet
}
\description{
Detect isoform based on the sample information file containing sample name,
tissue/treatment and FASTQ path
}
