% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_one_group}
\alias{detect_one_group}
\title{Detect Isoforms for a Tissue or Treatment Group}
\usage{
detect_one_group(
  sample_info,
  mirnas,
  max_indel_5p = 3,
  max_indel_3p = 3,
  max_snp_5p = 1,
  max_snp_3p = 1,
  max_snp_seed = 1,
  max_snp = 3,
  min_tpm = 1
)
}
\arguments{
\item{sample_info}{A data frame containing sample information for the
tissue/treatment group. Must include the following columns:
\itemize{
   \item name: Character vector of sample names.
   \item fq: Character vector of paths to FASTQ files.
}}

\item{mirnas}{A data frame containing reference miRNA information. Must
include the following columns:
\itemize{
  \item mature_id: Unique identifier of the mature miRNA.
  \item mature_seq: Sequence of the mature miRNA.
  \item seed_seq: Seed sequence.
  \item template_seq: Template sequence including flanking regions.
  \item flank_5p_seq: 5' flanking sequence of the mature miRNA.
  \item flank_3p_seq: 3' flanking sequence of the mature miRNA.
}}

\item{max_indel_5p}{An integer specifying the maximum allowed indels at
the 5' end. Must be non-negative.}

\item{max_indel_3p}{An integer specifying the maximum allowed indels at the
3' end. Must be non-negative.}

\item{max_snp_5p}{An integer specifying the maximum allowed mismatches at the
5' end. Must be non-negative.}

\item{max_snp_3p}{An integer specifying the maximum allowed mismatches at the
3' end. Must be non-negative.}

\item{max_snp_seed}{An integer specifying the maximum allowed mismatches in
the seed region. Must be non-negative.}

\item{max_snp}{An integer specifying the maximum allowed mismatches. Must be
non-negative.}

\item{min_tpm}{A numeric value specifying the minimum TPM threshold for
filtering low-expression isoforms. Isoforms with TPM below this threshold
are excluded.}
}
\value{
A list of data frames, one for each sample. Each data frame contains
  the following columns:
  \itemize{
    \item mature_id: Reference miRNA ID.
    \item mature_seq: Reference miRNA sequence.
    \item template_seq: Template sequence including flanking regions.
    \item read_seq: Read sequence matched to the reference.
    \item tpm: Numeric vector of TPM values.
    \item indel_5p: Indels at the 5' end.
    \item indel_3p: Indels at the 3' end.
    \item snp_5p: Mismatches at the 5' end.
    \item snp_3p: Mismatches at the 3' end.
    \item snp_seed: Mismatches in the seed.
    \item dist: Total edit distance.
    \item id: A unique ID to each isoform.
  }
}
\description{
This function detects miRNA isoforms (isomiRs) for all samples within a
tissue or treatment group. It processes each sample's FASTQ file, identifies
isomiRs, calculates TPM, filters low-expression isoforms.
}
\details{
The function performs the following steps for each sample:
1. Detects isomiRs using `detect_one_sample`.
2. Assigns a unique ID to each isoform (`mature_id:read_seq`).
3. Filters isoforms with TPM below `min_tpm`.
4. Ensures consistency across samples by retaining only isoforms present in
   all samples.
}
