% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_tissue}
\alias{detect_tissue}
\title{Detect isoforms from a tissue consist of several duplicates}
\usage{
detect_tissue(sample_info, mirnas, max_ed_5p, max_ed_3p, min_tpm)
}
\arguments{
\item{sample_info}{A data.frame with sample name, fq and group information
for replicates in a tissu/treatment}

\item{mirnas}{A data.frame storing mature miRNA ID, mature miRNA sequence,
mature start, seed on miRNA, precursor ID and precursor sequence}

\item{max_ed_5p}{the maximum distance between 5’ region of reads with
reference miRNA}

\item{max_ed_3p}{the maximum distance between 3' region of reads with
reference miRNA}

\item{min_tpm}{a isomiR is defined as “expressed” only when the TPM of all
biological replicates were greater than or equal to the threshold set}
}
\value{
a list
}
\description{
Detect isoforms from a tissue consist of several duplicates
}
