% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detectIsomirs}
\alias{detectIsomirs}
\title{Detect isoform based on the sample information file containing sample name,
tissue/treatment and FASTQ path}
\usage{
detectIsomirs(
  sampleInfoFile,
  spe,
  wordSize = 13,
  maxEd5p = 2,
  maxEd3p = 3,
  minTpm = 2
)
}
\arguments{
\item{sampleInfoFile}{description}

\item{spe}{description}

\item{wordSize}{description}

\item{maxEd5p}{description}

\item{maxEd3p}{description}

\item{minTpm}{description}
}
\value{
IsomirDataSet
}
\description{
Detect isoform based on the sample information file containing sample name,
tissue/treatment and FASTQ path
}
