% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detectOneTissue}
\alias{detectOneTissue}
\title{====================== private functions ====================================
Detect isoforms from a tissue consist of several duplicates}
\usage{
detectOneTissue(sampleInfo, mirnas, maxEd5p, maxEd3p, minTpm)
}
\arguments{
\item{sampleInfo}{A data.frame with sample name, fq and group information
for replicates in a tissu/treatment}

\item{mirnas}{A data.frame storing mature miRNA ID, mature miRNA sequence,
mature start, seed on miRNA, precursor ID and precursor sequence}

\item{maxEd5p}{the maximum distance between 5’ region of reads with
reference miRNA}

\item{maxEd3p}{the maximum distance between 3' region of reads with
reference miRNA}

\item{minTpm}{a isomiR is defined as “expressed” only when the TPM of all
biological replicates were greater than or equal to the threshold set}
}
\value{
a data.frame
}
\description{
====================== private functions ====================================
Detect isoforms from a tissue consist of several duplicates
}
