% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Cluster and visualize expression profile}
\usage{
draw_heatmap(expr, scale = "row")
}
\arguments{
\item{expr}{description}

\item{scale}{description}
}
\description{
Cluster and visualize expression profile
}
