% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{editDist}
\alias{editDist}
\title{Find the minimum number of edits to convert ‘s1‘ into ‘s2‘.}
\usage{
editDist(s1, s2)
}
\arguments{
\item{s1}{first string}

\item{s2}{second string}
}
\value{
edit distance
}
\description{
Find the minimum number of edits to convert ‘s1‘ into ‘s2‘.
}
