% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_isoforms}
\alias{find_isoforms}
\title{Find isoform from reads based on reference miRNAs}
\usage{
find_isoforms(mirnas, reads, max_ed_5p, max_ed_3p)
}
\arguments{
\item{mirnas}{A data.frame storing mature miRNA ID (matrue_id),
mature miRNA sequence (mature_seq), mature start (mature_start),
seed on miRNA (seed_seq), precursor ID (pre_id) and precursor sequence
(pre_seq)}

\item{reads}{A data.frame containing sequence and it's amount}

\item{max_ed_5p}{the maximum distance between 5’ region of reads with reference miRNA}

\item{max_ed_3p}{the maximum distance between 3' region of reads with reference miRNA}
}
\value{
a data.frame with columns: mature_id, read_seq, read_num and dist
}
\description{
Find isoform from reads based on reference miRNAs
}
