% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_cigar}
\alias{generate_cigar}
\title{Extract CIGAR of each sample replicate from isomiRs}
\usage{
generate_cigar(isomirs)
}
\arguments{
\item{isomirs}{isomiRs of all samples}
}
\value{
a data.frame
}
\description{
Extract CIGAR of each sample replicate from isomiRs
}
