% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_group_expr}
\alias{generate_group_expr}
\title{Extract expression profile of each sample replicate from isomiRs}
\usage{
generate_group_expr(g2isoforms)
}
\arguments{
\item{g2isoforms}{isomiRs of all samples}
}
\value{
a data.frame of expression
}
\description{
Extract expression profile of each sample replicate from isomiRs
}
