% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_isomir}
\alias{generate_isomir}
\title{Extract isomiRs from isoforms}
\usage{
generate_isomir(isoforms)
}
\arguments{
\item{isoforms}{description}
}
\value{
description
}
\description{
Extract isomiRs from isoforms
}
