% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_sample_expr}
\alias{generate_sample_expr}
\title{Generate Expression Matrix for Samples}
\usage{
generate_sample_expr(group2sample_isoforms)
}
\arguments{
\item{group2sample_isoforms}{A list of lists, where each inner list contains
isoform data frame for a specific group and sample. Each data frame must
include the following columns:
\itemize{
 \item read_seq: Character vector of isoform sequences.
 \item TPM: Numeric vector of TPM.
}}
}
\value{
A numeric data frame where:
  \itemize{
    \item Rownames: Isoform sequences (read_seq).
    \item Columns: Sample names.
    \item Values: TPM. Missing values are replaced with 0.
  }
}
\description{
This function generates an expression matrix for individual samples based on
isoform read TPM. It combines read TPM for each isoform across samples into
a single matrix, with isoforms as rows and samples as columns.
}
