% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generateExpr}
\alias{generateExpr}
\title{Extract expression profile of each sample replicate from isomiRs}
\usage{
generateExpr(isoformList)
}
\arguments{
\item{isomirList}{isomiRs of all samples}
}
\value{
a data.frame of expression
}
\description{
Extract expression profile of each sample replicate from isomiRs
}
