% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generateIsomirs}
\alias{generateIsomirs}
\title{Extract isomiRs from a list containing all isoforms}
\usage{
generateIsomirs(isoformList)
}
\arguments{
\item{isoformList}{a list containing all isoforms}
}
\value{
a list of all isomiRs
}
\description{
Extract isomiRs from a list containing all isoforms
}
