% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{get_alignment_by_ref}
\alias{get_alignment_by_ref}
\title{Extract Aligned Isoforms to a Reference Sequence}
\usage{
get_alignment_by_ref(x, ref)

\S4method{get_alignment_by_ref}{IsomirDataSet}(x, ref)
}
\arguments{
\item{x}{An object of class \linkS4class{IsomirDataSet} containing isomiR data.}

\item{ref}{A character string specifying the reference mature microRNA ID.}
}
\value{
A \linkS4class{DNAStringSet} object containing the aligned sequences
  of isomiRs and the reference template sequence.
}
\description{
This method extract the alignment of the sequences of isomiRs and their
reference template sequence using multiple sequence alignment (MSA). It is
useful for visualizing sequence variations and identifying potential editing
events in isomiRs.
}
\examples{
# Load example data
data(isomirs)

# Align isoforms for a specific reference microRNA
get_alignment_by_ref(isomirs, ref = "vvi-miR396c")

}
