% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{get_deg_data}
\alias{get_deg_data}
\title{Extract differential expression data from IsomirDataSet}
\usage{
get_deg_data(x, treatment, control)

\S4method{get_deg_data}{IsomirDataSet}(x, treatment, control)
}
\arguments{
\item{x}{An object of class \linkS4class{IsomirDataSet}}

\item{treatment}{Character vector specifying the treatment group name}

\item{control}{Character vector specifying the control group name}
}
\value{
A list containing two elements:
\itemize{
  \item \code{expr} - Expression matrix with sequences as rows and samples
        as columns. For type="ref", contains reference sequences;
        for type="isoform", contains isomiR sequences.
  \item \code{sample_info} - Data frame containing sample metadata with
        group factor ordered as c(treatment, control)
}
}
\description{
This function extracts expression data and sample information from an
IsomirDataSet object for specified treatment and control groups, preparing
it for differential expression analysis. The function allows selecting
either reference sequences or isoform sequences.
}
\examples{
data(isomirs)
deg_data <- get_deg_data(isomirs,
                        treatment = "Flower_FB",
                        control = "Inflorescence_WD")

}
