% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{get_expr_by_ref}
\alias{get_expr_by_ref}
\title{Extract Expression Data for IsomiRs of a Specific Reference}
\usage{
get_expr_by_ref(x, ref)

\S4method{get_expr_by_ref}{IsomirDataSet}(x, ref)
}
\arguments{
\item{x}{An object of class \linkS4class{IsomirDataSet} containing isomiR
data.}

\item{ref}{A character string specifying the reference mature microRNA ID.}
}
\value{
A data frame containing the expression data for isomiRs associated
  with the specified reference microRNA. The row names of the data frame are
  replaced with unique CIGAR strings for easier interpretation.
}
\description{
This method extracts expression data for isomiRs associated with a specified
reference mature microRNA. It is useful for analyzing the expression patterns
of isomiRs derived from a specific microRNA across different conditions or
groups.
}
\examples{
# Load example data
data(isomirs)

# Extract expression data for a specific reference microRNA
get_expr_by_ref(isomirs, ref = "vvi-miR396c")

}
