% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{get_isoform_num}
\alias{get_isoform_num}
\title{Count Isoforms per Reference miRNA}
\usage{
get_isoform_num(x)

\S4method{get_isoform_num}{IsomirDataSet}(x)
}
\arguments{
\item{x}{An \linkS4class{IsomirDataSet} object containing isomiR data.
Requires the \strong{ref2isomir} slot, which stores a list of reference
miRNAs with associated isoforms.}
}
\value{
A named numeric vector where:
  \itemize{
    \item Names: Reference miRNA IDs.
    \item Values: Number of isoforms associated with each reference miRNA.
  }
}
\description{
This method calculates the number of isoforms associated with each reference
miRNA in an \linkS4class{IsomirDataSet} object. It returns a vector of counts
for reference miRNAs.
}
\examples{
# Load example IsomirDataSet
data(isomirs)

# Count isoforms per reference miRNA
get_isoform_num(isomirs)

}
