% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/methods.R
\docType{methods}
\name{get_isomir_by_ref}
\alias{get_isomir_by_ref}
\title{Extract IsomiR Data for a Specific Reference miRNA}
\usage{
get_isomir_by_ref(x, ref)

\S4method{get_isomir_by_ref}{IsomirDataSet}(x, ref)
}
\arguments{
\item{x}{An \linkS4class{IsomirDataSet} object containing isomiR data.
Requires the \strong{ref2isomir} slot, which stores a list of reference
miRNAs with associated isomiR information.}

\item{ref}{A character string specifying the reference miRNA ID.}
}
\value{
A list containing the isomiR data for the specified reference miRNA.
}
\description{
This method extracts the isomiR data associated with a specified reference
miRNA from an \linkS4class{IsomirDataSet} object. It returns the isomiR
information stored in the \strong{ref2isomir} slot for the given reference
miRNA.
}
\examples{
# Load example IsomirDataSet
data(isomirs)

# Extract isomiR data for a specific reference miRNA
get_isomir_by_ref(isomirs, ref = "vvi-miR396c")

}
