% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_ref,IsomirDataSet-method}
\alias{get_ref,IsomirDataSet-method}
\title{=============================IsomirDataSet =================================}
\usage{
\S4method{get_ref}{IsomirDataSet}(x)
}
\description{
=============================IsomirDataSet =================================
}
