% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{group_expr}
\alias{group_expr}
\title{Extract expression profile of each tissue/treatment from isomiRs}
\usage{
group_expr(isomirs)
}
\arguments{
\item{isomirs}{isomiRs of all samples}
}
\value{
a data.frame
}
\description{
Extract expression profile of each tissue/treatment from isomiRs
}
