% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{has_isomir,IsomirDataSet-method}
\alias{has_isomir,IsomirDataSet-method}
\title{============================= IsomirDataSet ================================}
\usage{
\S4method{has_isomir}{IsomirDataSet}(x)
}
\description{
============================= IsomirDataSet ================================
}
