% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{Isomir-class}
\alias{Isomir-class}
\title{An S4 class to represent a isomiR}
\description{
An S4 class to represent a isomiR
}
\section{Slots}{

\describe{
\item{\code{mature_id}}{microRNA ID}

\item{\code{read_seqs}}{Vector of read sequences}

\item{\code{dist}}{vector of edit distance}

\item{\code{cigars}}{vector of CIGAR}
}}

