% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{IsomirDataSet-class}
\alias{IsomirDataSet-class}
\title{An S4 class to represent a dataset containing all the isomiR information
  detected from RNA-Seq samples.}
\description{
An S4 class to represent a dataset containing all the isomiR information
  detected from RNA-Seq samples.
}
\section{Slots}{

\describe{
\item{\code{isoformList}}{A list of isoforms}

\item{\code{expr}}{A data.fram of expression profile}

\item{\code{isomirList}}{A list of isomiRs}
}}

