% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_reads}
\alias{load_reads}
\title{Load reads in a FASTQ File}
\usage{
load_reads(fq_file)
}
\arguments{
\item{fq_file}{Path to the FASTQ file.}
}
\value{
A DataFrame containing two columns:
  - `read_seq`: The sequence of the read.
  - `tpm`: TPM of the read in the file.
}
\description{
This function reads a FASTQ file and marks duplicate reads
  based on their sequences.
}
\examples{
\dontrun{
  result <- load_reads("example.fq")
}
}
