% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_reads_gz}
\alias{load_reads_gz}
\title{Load reads in a GZIP-compressed FASTQ File}
\usage{
load_reads_gz(fq_file)
}
\arguments{
\item{fq_file}{Path to the GZIP-compressed FASTQ file.}
}
\value{
A DataFrame containing two columns:
  - `read_seq`: The sequence of the read.
  - `tpm`: TPM of the read in the file.
}
\description{
This function reads a GZIP-compressed FASTQ file and marks
  duplicate reads based on their sequences.
}
\examples{
\dontrun{
  result <- load_read_gz("example.fq.gz")
}
}
