% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBase.R
\name{loadMirna}
\alias{loadMirna}
\title{Load miRNAs of a species from mirbase.db}
\usage{
loadMirna(spe, wordSize = 13)
}
\arguments{
\item{spe}{species}

\item{word_size}{Length of seed in mature miRNA}
}
\value{
A data.frame storing mature miRNA ID (matrue_id),
  mature miRNA sequence (mature_seq), mature start (mature_start),
  seed on miRNA (seed_seq), precursor ID (pre_id) and precursor sequence
  (pre_seq)
}
\description{
Load miRNAs of a species from mirbase.db
}
\examples{
load_mirna("vvi", 13)

}
