% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{plot_expr}
\alias{plot_expr}
\title{Cluster and visualize expression profile}
\usage{
plot_expr(expr, scale = "row")
}
\arguments{
\item{expr}{description}

\item{scale}{description}
}
\description{
Cluster and visualize expression profile
}
