% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsomirDataSet.R
\name{plotExpr,IsomirDataSet-method}
\alias{plotExpr,IsomirDataSet-method}
\title{==================  Methods ======================
Cluster and visualize expression profile}
\usage{
\S4method{plotExpr}{IsomirDataSet}(x)
}
\arguments{
\item{expr}{description}
}
\description{
==================  Methods ======================
Cluster and visualize expression profile
}
