% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsomirDataSet.R
\name{plotExpr}
\alias{plotExpr}
\title{================================= Generics =================================}
\usage{
plotExpr(x)
}
\description{
================================= Generics =================================
}
