% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_isomirs}
\alias{detect_isomirs}
\title{Detect IsomiRs from Small RNA-Seq Data}
\usage{
detect_isomirs(
  sample_info_file,
  fq_dir,
  mirnas,
  max_indel_5p = 3,
  max_indel_3p = 3,
  max_snp_5p = 1,
  max_snp_3p = 1,
  max_snp_seed = 1,
  max_snp = 3,
  min_tpm = 1
)
}
\arguments{
\item{sample_info_file}{A character string specifying the path to the sample
information file. The file should be a CSV file with columns: name, group,
and fq (FASTQ file name).}

\item{fq_dir}{A character string specifying the directory containing the
FASTQ files.}

\item{mirnas}{A data frame containing microRNA data, typically generated by
\code{load_mirnas}}

\item{max_indel_5p}{An integer specifying the maximum allowed indels at
the 5' end. Must be non-negative (default: 3).}

\item{max_indel_3p}{An integer specifying the maximum allowed indels at the
3' end. Must be non-negative (default: 3).}

\item{max_snp_5p}{An integer specifying the maximum allowed mismatches at the
5' end. Must be non-negative (default: 1).}

\item{max_snp_3p}{An integer specifying the maximum allowed mismatches at the
3' end. Must be non-negative (default: 1).}

\item{max_snp_seed}{An integer specifying the maximum allowed mismatches in
the seed region. Must be non-negative (default: 1).}

\item{max_snp}{An integer specifying the maximum allowed mismatches. Must be
non-negative (default: 3).}

\item{min_tpm}{A numeric value specifying the minimum TPM threshold for
filtering low-expression isomiRs (default: 1).}
}
\value{
An object of class \linkS4class{IsomirDataSet}.
}
\description{
This function detects isomiRs from small RNA-Seq data for multiple samples.
It processes FASTQ files, identifies isomiRs, and generates expression
profiles for samples and groups. The results are returned as an
\linkS4class{IsomirDataSet} object, which can be used for further analysis.
}
\details{
This function performs the following steps:
1. Reads the sample information file and validates the input data (e.g.,
   checks for duplicate sample names or missing FASTQ files).
2. Groups samples by their experimental conditions (e.g., tissue or treatment).
3. Detects isomiRs for each sample using the \code{\link{detect_one_tissue}}
   function.
4. Aggregates isomiR expression data across replicates within each group.
5. Generates expression profiles for individual samples and groups.
6. Clusters isomiRs based on reference mature microRNAs.
7. Returns an \linkS4class{IsomirDataSet} object for further analysis.
}
