% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{detect_one_sample}
\alias{detect_one_sample}
\title{Detect Isoforms from a Single Small RNA-Seq Sample}
\usage{
detect_one_sample(
  fq_file,
  mirnas,
  max_indel_5p = 3,
  max_indel_3p = 3,
  max_snp_5p = 1,
  max_snp_3p = 1,
  max_snp_seed = 1,
  max_snp = 3
)
}
\arguments{
\item{fq_file}{A character string specifying the path to the FASTQ file. The
file can be compressed (`.gz`) or uncompressed.}

\item{mirnas}{A data frame containing reference miRNA information. Must
include the following columns:
  \itemize{
    \item mature_id: Unique identifier of the mature miRNA.
    \item mature_seq: Sequence of the mature miRNA.
    \item seed_seq: Seed sequence (center region of the mature miRNA).
    \item template_seq: Template sequence including flanking regions.
    \item flank_5p_seq: 5' flanking sequence of the mature miRNA.
    \item flank_3p_seq: 3' flanking sequence of the mature miRNA.
  }}

\item{max_indel_5p}{An integer specifying the maximum allowed indels at
the 5' end. Must be non-negative.}

\item{max_indel_3p}{An integer specifying the maximum allowed indels at the
3' end. Must be non-negative.}

\item{max_snp_5p}{An integer specifying the maximum allowed mismatches at the
5' end. Must be non-negative.}

\item{max_snp_3p}{An integer specifying the maximum allowed mismatches at the
3' end. Must be non-negative.}

\item{max_snp_seed}{An integer specifying the maximum allowed mismatches in
the seed region. Must be non-negative.}

\item{max_snp}{An integer specifying the maximum allowed mismatches. Must be
non-negative.}
}
\value{
A data frame containing detected isoforms with the following columns:
  \itemize{
    \item mature_id: Reference miRNA ID.
    \item mature_seq: Reference miRNA sequence.
    \item template_seq: Template sequence including flanking regions.
    \item read_seq: Read sequence matched to the reference.
    \item tpm: Numeric vector of TPM values.
    \item indel_5p: Indels at the 5' end.
    \item indel_3p: Indels at the 3' end.
    \item snp_5p: Mismatches at the 5' end.
    \item snp_3p: Mismatches at the 3' end.
    \item snp_seed: Mismatches in the seed.
    \item dist: Total edit distance.
  }
}
\description{
This function detects miRNA isoforms (isomiRs) from a single small RNA-Seq
sample. It reads the sequencing data from a FASTQ file, caculate reads TPM,
and identifies isomiRs by comparing reads to reference miRNAs with specified
InDel and SNP thresholds.
}
\details{
The function performs the following steps:
1. Checks if max_indel_5p, max_indel_3p, max_snp_5p, max_snp_3p and max_snp
   are non-negative.
2. Reads the FASTQ file and caculate reads TPM using \code{load_reads} or
   \code{load_reads_gz}.
3. Identifies isomiRs by comparing reads to reference miRNAs using the
   \code{find_isoforms} function.
}
