% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_isoforms}
\alias{find_isoforms}
\title{Find isoform from reads based on reference miRNAs}
\usage{
find_isoforms(mirnas, reads, max_ed_5p, max_ed_3p)
}
\arguments{
\item{mirnas}{a data.frame storing mature miRNA ID (matrue_ID),
mature miRNA sequence (mature_seq), mature start (mature_start),
seed on miRNA (seed_seq), precursor ID (pre_ID) and precursor sequence
(pre_seq)}

\item{reads}{a data.frame containing sequence and it's amount}

\item{maxEd5p}{the maximum distance between 5’ region of reads with
reference miRNA}

\item{maxEd3p}{the maximum distance between 3' region of reads with
reference miRNA}
}
\value{
a data.frame with columns: mature_ID, read_seq, read_num and dist
}
\description{
Find isoform from reads based on reference miRNAs
}
