% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_isoforms}
\alias{find_isoforms}
\title{Find Isoforms}
\usage{
find_isoforms(
  mirnas,
  reads,
  max_indel_5p,
  max_indel_3p,
  max_snp_5p,
  max_snp_3p,
  max_snp_seed,
  max_snp
)
}
\arguments{
\item{mirnas}{A DataFrame containing miRNA information. The DataFrame should
include the following columns:
\itemize{
  \item mature_id: Unique identifier of the mature miRNA.
  \item mature_seq: Sequence of the mature miRNA.
  \item seed_seq: Seed sequence of the mature miRNA.
  \item template_seq: Template sequence including flanking regions.
  \item flank_5p_seq: 5' flanking sequence of the mature miRNA.
  \item flank_3p_seq: 3' flanking sequence of the mature miRNA.
}}

\item{reads}{A DataFrame containing read information. The DataFrame should
include the following columns:
\itemize{
  \item read_seq: Sequence of the read.
  \item tpm: TPM of the read.
}}

\item{max_indel_5p}{An integer specifying the maximum allowed indels at the
5' end. Must be non-negative.}

\item{max_indel_3p}{An integer specifying the maximum allowed indels at the
3' end. Must be non-negative.}

\item{max_snp_5p}{An integer specifying the maximum allowed mismatches at
the 5' end. Must be non-negative.}

\item{max_snp_3p}{An integer specifying the maximum allowed mismatches at
the 3' end. Must be non-negative.}

\item{max_snp_seed}{An integer specifying the maximum allowed mismatches in
the seed region. Must be non-negative.}

\item{max_snp}{An integer specifying the maximum allowed mismatches. Must be
non-negative.}
}
\value{
A DataFrame containing the detected isoforms with the following columns:
  \itemize{
    \item mature_id: Reference miRNA ID.
    \item mature_seq: Reference miRNA sequence.
    \item seed_seq: Seed sequence of the mature miRNA.
    \item template_seq: Template sequence including flanking regions.
    \item read_seq: Read sequence matched to the reference.
    \item read_num: Abundance of the read.
    \item tpm: TPM of the read.
    \item indel_5p: Indels at the 5' end.
    \item indel_3p: Indels at the 3' end.
    \item snp_5p: Mismatches at the 5' end.
    \item snp_3p: Mismatches at the 3' end.
    \item snp_seed: Mismatches in the seed.
    \item dist: Total edit distance.
  }
}
\description{
Find isoforms for given miRNAs and reads.
}
\details{
This function compares read sequences to reference miRNAs and identifies
isoforms based on the specified thresholds for indels and mismatches. It calculates the edit distance between the read sequences and the reference miRNAs, and returns a DataFrame containing the detected isoforms that meet the criteria.
}
\examples{
\dontrun{
}
}
