% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_expr}
\alias{generate_expr}
\title{Extract expression profile of each sample replicate from isomiRs}
\usage{
generate_expr(isoform_list)
}
\arguments{
\item{isomir_list}{isomiRs of all samples}
}
\value{
a data.frame of expression
}
\description{
Extract expression profile of each sample replicate from isomiRs
}
