% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_group_expr}
\alias{generate_group_expr}
\title{Generate Expression Matrix for Groups}
\usage{
generate_group_expr(group2isoforms)
}
\arguments{
\item{group2isoforms}{A list of data frames, where each data frame contains
isoform data for a specific group. Each data frame must include the
following columns:
\itemize{
  \item read_seq: Character vector of isoform sequences.
  \item tpm: Numeric vector of mean TPM values across all the sampel in the
        a group.
}}
}
\value{
A numeric data frame where:
 \itemize{
   \item Rownames: Isoform sequences (read_seq).
   \item Columns: Group names (names of group2isoforms).
   \item Values: TPM values. Missing values are replaced with 0.
 }
}
\description{
This function generates an expression matrix for groups (e.g., tissues or
treatments) based on isoform expression data. It combines TPM
(Transcripts Per Million) values for each isoform across groups into a single
matrix, with isoforms as rows and groups as columns.
}
