% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risoma.R
\name{generate_isomir}
\alias{generate_isomir}
\title{Extract isomiRs from a list containing all isoforms}
\usage{
generate_isomir(all_isoforms)
}
\arguments{
\item{all_isoforms}{a list containing all isoforms}
}
\value{
a list of all isomiRs
}
\description{
Extract isomiRs from a list containing all isoforms
}
