% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{IsomirDataSet-class}
\alias{IsomirDataSet-class}
\title{An S4 class to represent a dataset containing all the isomiR information
detected from RNA-Seq samples.}
\description{
An S4 class to represent a dataset containing all the isomiR information
detected from RNA-Seq samples.
}
\section{Slots}{

\describe{
\item{\code{sample_info}}{A data.frame containing sample information}

\item{\code{group2isoforms}}{A list of isoforms grouped by sample groups}

\item{\code{sample_expr}}{A data.frame of expression profile for samples}

\item{\code{group_expr}}{A data.frame of expression profile for groups}

\item{\code{ref2isomir}}{A list of isomiRs grouped by reference mature microRNAs}
}}

