% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirbase.R
\name{load_mirnas}
\alias{load_mirnas}
\title{Load microRNA Data from miRBase}
\usage{
load_mirnas(spe, flank_5p_len = 3, flank_3p_len = 3, seed_len = 16)
}
\arguments{
\item{spe}{A character string specifying the species name (e.g., "ath" for
Arabidopsis thaliana).}

\item{flank_5p_len}{An integer specifying the length of the 5' flanking
region (default: 3).}

\item{flank_3p_len}{An integer specifying the length of the 3' flanking
region (default: 3).}

\item{seed_len}{Lenght of the seed sequence. (default: 16)}
}
\value{
A data frame containing the following columns:
 \itemize{
   \item mature_id: The ID of the mature miRNA.
   \item mature_seq: The sequence of the mature miRNA.
   \item mature_start: The start position of the mature miRNA in the precursor sequence.
   \item mature_end: The end position of the mature miRNA in the precursor sequence.
   \item seed_seq: The seed sequence of the mature miRNA.
   \item pre_id: The ID of the precursor miRNA.
   \item template_seq: The template sequence, including the mature miRNA and its flanking regions.
   \item flank_5p_seq: The 5' flanking sequence.
   \item flank_3p_seq: The 3' flanking sequence.
 }
}
\description{
This function retrieves microRNA (miRNA) data from the miRBase database for a
specified species. It extracts mature miRNA sequences, seed sequences, and 
flanking regions, and returns the data in a structured data frame.
}
\examples{
# Load miRNA data for Arabidopsis thaliana
mirnas <- load_mirnas(spe = "ath")

# View the first few rows of the data
head(mirnas)

}
